Add better /info routing
Use /info/bankett instead of the shitty /info?side=bankett

The old url will 301 Moved Permanently to the new url
