+++ components/CompactProgram/index.js
@@ -70,14 +70,14 @@ const CompactProgram = () => (
         </Link>
       </Tile>
       <Tile color={itdageneLightBlue}>
-        <Link href="/info?side=stands">
+        <Link href="/info/stands">
           <CenterIt text>
             <StyledLink>Stands</StyledLink>
           </CenterIt>
         </Link>
       </Tile>
       <Tile color={itdageneYellow}>
-        <Link href="/info?side=bankett">
+        <Link href="/info/bankett">
           <CenterIt text>
             <StyledLink>Bankett</StyledLink>
           </CenterIt>
