+++ components/Companies/Companies.js
@@ -38,7 +38,7 @@ const Companies = ({ query }: Props) => {
       </Flex>
       <h4 style={{ textAlign: 'center' }}>
         <Link>
-          <a href="/info?side=stands"> Kart over stands </a>
+          <a href="/info/stands"> Kart over stands </a>
         </Link>
       </h4>
     </Fragment>
